fix(trusty-search): probe deeper index path for TCC detection + surface leaked-probe-thread count (review #727 findings 2,3)
Finding 2 (medium): probe_volume now receives a `probe_path` (the actual sample
index directory inside the volume, e.g. /Volumes/SSD1/Projects/myrepo) in
addition to `volume_root` (retained for logging). The OS thread calls
`stat(probe_path)`, not `stat(volume_root)`. On macOS, `stat` on the volume
mount-point root can succeed even when TCC denies inner-file access; probing
the deeper representative index path is what actually detects the
TCC-blocked-inside-volume scenario issue #723 targets. probe_all_volumes passes
the collected sample path as the probe target to probe_volume. Once-per-volume
design preserved.

Finding 3 (medium): process-global `LEAKED_PROBE_THREAD_COUNT: AtomicUsize`
tracks probe threads abandoned on deadline. probe_volume increments it and
emits `tracing::warn!` with the running total on every timeout. The counter is
exposed via `leaked_probe_thread_count()` → re-exported from warm_boot → read
by `GET /health` as `warmboot_leaked_probe_threads` (integer, always present).

Tests added: `probe_uses_sample_path_not_volume_root`,
`probe_timeout_increments_leaked_thread_count`,
`health_includes_warmboot_leaked_probe_threads`. Tests split to
`probe_tests.rs` to keep probe.rs under the 500-line cap.
