+++ .line-cap-allowlist.tsv
@@ -142,7 +142,7 @@ crates/trusty-search/src/commands/doctor_checks.rs	612
 crates/trusty-search/src/commands/integrate.rs	688
 crates/trusty-search/src/commands/reindex_engine.rs	1438
 crates/trusty-search/src/commands/reindex_ui.rs	898
-crates/trusty-search/src/commands/start.rs	2139
+crates/trusty-search/src/commands/start.rs	2137
 crates/trusty-search/src/core/chunker/mod.rs	1883
 crates/trusty-search/src/core/classifier.rs	1032
 crates/trusty-search/src/core/corpus.rs	1420
@@ -167,7 +167,7 @@ crates/trusty-search/src/service/embedder_supervisor.rs	947
 crates/trusty-search/src/service/grep.rs	827
 crates/trusty-search/src/service/persistence.rs	990
 crates/trusty-search/src/service/reindex/mod.rs	3669
-crates/trusty-search/src/service/server.rs	5493
+crates/trusty-search/src/service/server.rs	5543
 crates/trusty-search/src/service/walker.rs	1343
 crates/trusty-search/tests/baseline_trusty_tools.rs	721
 crates/trusty-search/tests/benchmark_harness.rs	740
